+++ gamemodes/terrortown/gamemode/shared/lang/italian.lua
@@ -1315,3 +1315,6 @@ L.corpse_hint_inspect_only_credits = "Premi [{usekey}] per ricevere i crediti. S
 --L.pickup_error_spec = "You cannot pickup this as a spectator."
 --L.pickup_error_owns = "You cannot pickup this because you already own the same weapon."
 --L.pickup_error_noslot = "You cannot pickup this because you have no free slot available."
+
+-- 2020-08-11
+--L.f1_settings_shop_desc_double_click = "Enable to buy an item in the shop by double clicking on it."