+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Added
 
+- Added ConVar to toggle double-click buying
+
+### Fixed
+
+- Fixed foregoing fetch fix
+
+## [v0.7.3b](https://github.com/TTT-2/TTT2/tree/v0.7.3b) (2020-08-09)
+
+### Added
+
 - Added a new custom file loader that loads lua files from `lua/terrortown/autorun/`
   - it basically works the same as the native file loader
   - there are three subfolders: `client`, `server` and `shared`